Add discovery of matched reader
