+++ dds/src/implementation/domain_participant_backend/entities/domain_participant.rs
@@ -1160,78 +1160,6 @@ impl DomainParticipantEntity {
             && self.user_defined_subscriber_list.is_empty()
             && no_user_defined_topics
     }
-
-    pub fn process_discovered_readers(&mut self) {
-        for subscription_builtin_topic_data in self.discovered_reader_list.values() {
-            for publisher in &mut self.user_defined_publisher_list {
-                let is_any_name_matched = subscription_builtin_topic_data
-                    .partition
-                    .name
-                    .iter()
-                    .any(|n| publisher.qos().partition.name.contains(n));
-
-                let is_any_received_regex_matched_with_partition_qos =
-                    subscription_builtin_topic_data
-                        .partition
-                        .name
-                        .iter()
-                        .filter_map(|n| glob_to_regex(n).ok())
-                        .any(|regex| {
-                            publisher
-                                .qos()
-                                .partition
-                                .name
-                                .iter()
-                                .any(|n| regex.is_match(n))
-                        });
-
-                let is_any_local_regex_matched_with_received_partition_qos = publisher
-                    .qos()
-                    .partition
-                    .name
-                    .iter()
-                    .filter_map(|n| glob_to_regex(n).ok())
-                    .any(|regex| {
-                        subscription_builtin_topic_data
-                            .partition
-                            .name
-                            .iter()
-                            .any(|n| regex.is_match(n))
-                    });
-
-                let is_partition_matched = subscription_builtin_topic_data.partition
-                    == publisher.qos().partition
-                    || is_any_name_matched
-                    || is_any_received_regex_matched_with_partition_qos
-                    || is_any_local_regex_matched_with_received_partition_qos;
-                if is_partition_matched {
-                    let publisher_qos = publisher.qos().clone();
-                    for data_writer in publisher.data_writer_list_mut() {
-                        let is_matched_topic_name = subscription_builtin_topic_data.topic_name()
-                            == data_writer.topic_name();
-                        let is_matched_type_name = subscription_builtin_topic_data.get_type_name()
-                            == data_writer.type_name();
-
-                        if is_matched_topic_name && is_matched_type_name {
-                            let incompatible_qos_policy_list =
-                                get_discovered_reader_incompatible_qos_policy_list(
-                                    &data_writer.qos(),
-                                    &subscription_builtin_topic_data,
-                                    &publisher_qos,
-                                );
-                            if incompatible_qos_policy_list.is_empty() {
-                                data_writer.add_matched_subscription(
-                                    subscription_builtin_topic_data.clone(),
-                                );
-                            } else {
-                                todo!("Incompatible reader found")
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
 }
 
 fn get_discovered_writer_incompatible_qos_policy_list(
